modify the table and migration to use numeric and integer

+++ backend/db/migrations/20241101225450-create-review.js
@@ -38,7 +38,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.NUMERIC(2, 1),
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       createdAt: {
